# api/openapi.yaml
openapi: 3.1.0
info:
  title: TaskForge API
  version: 0.3.0-m3
  description: |
    Durable job ingress plus internal worker registration, session heartbeat,
    atomic claim, fenced lease renewal, and fenced start/success operations.

    This document describes **only the endpoints that are implemented**. The rest
    of the V1 surface listed in `docs/PROJECT_SPEC.md` — job listing,
    cancellation, retry, DLQ, workers, queues — arrives with the milestone that
    makes it real. An endpoint is never documented here before it works.

    ## Authentication

    None yet. Milestone M3 attributes every request to a single configured
    development scope, and the service therefore binds to loopback only.
    Database-backed API keys are milestone M5. See `docs/CURRENT_STATE.md`.

    Routes under `/internal/v1` are for the trusted, compiled-in worker process.
    They are not a public user API and must not be exposed off loopback before
    worker/control scopes land in M5.

    ## Idempotency

    `POST /v1/jobs` requires an `Idempotency-Key` header. TaskForge canonicalizes
    every job-defining field and hashes it into a request fingerprint:

    - first use of a key → `201 Created`
    - same key, equivalent request → `200 OK` with the original job
    - same key, different request → `409 Conflict`

    "Equivalent" means equivalent after normalization: reordering JSON keys,
    stating a default explicitly, or repeating a required capability all describe
...
  license:
...
servers:
...
paths:
  /v1/jobs:
    post:
      tags: [Jobs]
      summary: Submit a job
      description: |
        Durably accepts a job. The job, its idempotency record, and the broker
        notification are written in one PostgreSQL transaction, so a success
        response means the job is durable.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
          description: Caller-generated key, unique per distinct submission.
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 128 }
          description: |
            Echoed back and included in error bodies. Ignored if it is longer
            than 128 bytes or contains control characters.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitJobRequest' }
            examples:
              minimal:
                value:
                  queue: default
                  job_type: demo.echo
                  payload: { message: hello }
              full:
                value:
                  queue: default
                  job_type: demo.sleep
                  payload: { duration_ms: 5000 }
                  priority: 70
                  max_attempts: 5
                  timeout_seconds: 45
                  scheduled_at: null
                  required_capabilities: [cpu]
      responses:
        '201':
          description: The job was created.
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '200':
          description: An identical earlier submission already created this job.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '400':
          description: |
            The body was not a single valid JSON object, or it contained an
            unknown field. Unknown fields are rejected rather than ignored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: This `Idempotency-Key` was already used for a different request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: The body exceeded the configured size limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: |
            The body was valid JSON but the request was invalid — including a
            non-null `scheduled_at`, which is not implemented in this milestone.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                scheduledNotImplemented:
                  value:
                    error:
                      code: validation_failed
                      message: the request was rejected by validation
                      details:
                        - field: scheduled_at
                          message: delayed execution is not implemented in this milestone; send null or omit the field
                      request_id: 6a1f...
        '500':
          description: Internal error. The response is sanitized; use `request_id` to find the cause in the logs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Read a job
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The job.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404':
          description: |
            No such job in the caller's scope. A malformed id also answers 404,
            so a caller cannot distinguish an invalid id from someone else's.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /internal/v1/worker-sessions/{worker_session_id}:
    put:
      tags: [Worker Control]
      summary: Idempotently register one worker process session
      description: |
        `worker_session_id` is generated once per process boot and reused when
        an ambiguous registration response is retried. Reusing it with different
        immutable registration data conflicts. A new session for the same stable
        worker name fences the prior session; old active leases remain durable
        and capacity-consuming until they expire and the reconciler abandons
        their attempts.
      parameters:
        - name: worker_session_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkerSessionRegistration' }
      responses:
        '200':
          description: The registered or identically replayed process session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerSession' }
        '400':
          description: Malformed JSON or an unknown field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: The session id conflicts or the session was already replaced.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: The body exceeded the configured size limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Invalid registration data.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Sanitized internal error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: |
            A database call in this request failed because the request's own
            server-owned deadline elapsed, so the durable outcome is unknown to
            the caller. Code `service_unavailable`.

            The deadline can elapse while acquiring a lock, while executing a
            statement, or during COMMIT. A COMMIT that is cut short is
            ambiguous, so this response never asserts that nothing was
            committed.

            **Retry:** repeat the identical registration request on the same
            `worker_session_id` path with the same immutable registration body.
            Registration is idempotent under that session identity: a replay
            that matches returns the already-registered session rather than
            creating or fencing a second one.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: service_unavailable
                  message: >-
                    the request exceeded its deadline before the durable outcome was
                    known; retry the identical request, which is safe because every
                    worker-control operation is idempotent under the identity and
                    fencing identifiers it already carries
  /internal/v1/worker-sessions/{worker_session_id}/heartbeat:
    post:
      tags: [Worker Control]
      summary: Report that one process session is alive
      description: |
        Advances the session's `last_heartbeat_at` using PostgreSQL receipt time,
        sampled after the session row is locked.

        The request deliberately carries no timestamp. Worker-supplied wall-clock
        time is never authoritative for staleness, so the response echoes the
        server time that was actually accepted and a worker confirms liveness
        from that rather than from its own clock.

        Only a current `HEALTHY` session is accepted. A replaced, unhealthy, or
        ended session answers `409 worker_session_unavailable` and is never
        returned to health; that process must establish a new boot session.

        Repeating a heartbeat is harmless. It may advance the stored timestamp
        again — that is what a heartbeat is for — but it can never create a
        session or revive a fenced one.
      parameters:
        - name: worker_session_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HeartbeatRequest' }
      responses:
        '200':
          description: The accepted PostgreSQL receipt time for this session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Heartbeat' }
        '400':
          description: Malformed JSON or an unknown field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: |
            The session is not current and healthy. Code
            `worker_session_unavailable`. A late heartbeat never revives a
            fenced session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: The body exceeded the configured size limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: An identifier was invalid.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Sanitized internal error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: |
            A database call in this request failed because the request's own
            server-owned deadline elapsed, so the durable outcome is unknown to
            the caller. Code `service_unavailable`.

            The deadline can elapse while acquiring a lock, while executing a
            statement, or during COMMIT. A COMMIT that is cut short is
            ambiguous, so this response never asserts that nothing was
            committed.

            **Retry:** repeat the identical request — the same
            `worker_session_id` in the path and the same `worker_id` in the
            body. A heartbeat is idempotent under that session identity: a
            replay may advance the accepted receipt time again, which is
            exactly what a heartbeat does, and can never create a session or
            revive one that was already fenced.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: service_unavailable
                  message: >-
                    the request exceeded its deadline before the durable outcome was
                    known; retry the identical request, which is safe because every
                    worker-control operation is idempotent under the identity and
                    fencing identifiers it already carries
  /internal/v1/leases/{lease_id}/renew:
    post:
      tags: [Worker Control]
      summary: Fenced, idempotent lease renewal
      description: |
        Extends one lease's authority window so cooperative work may safely run
        across more than one lease. It extends lease authority **only**: the
        job's overall `timeout_seconds` budget is measured once from execution
        start and renewal never resets it.

        Renewal is fenced by the complete five-part fence — `lease_id` in the
        path plus `job_id`, `attempt_id`, `worker_id`, and `worker_session_id` —
        and additionally by a client-generated `renewal_request_id` and the
        `expected_renewal_version` the caller believes is current. That pair is
        what makes renewal safe to retry rather than silently doubling
        authority:

        - an exact replay of the renewal that produced the current generation
          returns the stored window unchanged and moves nothing;
        - a delayed older generation, or a second distinct request for the same
          generation, mutates nothing and returns `409 renewal_conflict`;
        - reusing a `renewal_request_id` that is currently recorded on a different
          lease returns the same stable `renewal_conflict`, never a leaked
          uniqueness error.

        Only each lease's current renewal identity is retained, so that last rule
        constrains identities that are in force, not every identity ever used.
        Once a lease renews again its previous id is free, and reusing it on
        another lease renews that lease normally. Nothing is weakened by this: an
        identity authorizes nothing on its own, and whether a lease is extended is
        decided by its own fence and `expected_renewal_version`.

        Renewal never resurrects an expired, completed, released, or reconciled
        lease. PostgreSQL time is resampled after every authority lock, so a
        renewal that waited across the expiry boundary is rejected with
        `409 lease_expired` rather than accepted on a stale sample.

        `lease_remaining_milliseconds` is measured by PostgreSQL after those
        locks. A worker converts it into a conservative monotonic deadline
        instead of comparing its own wall clock with `lease_expires_at`.
      parameters:
        - name: lease_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenewLeaseRequest' }
      responses:
        '200':
          description: The committed renewal window, or an exact replay of it.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Renewal' }
        '400':
          description: Malformed JSON or an unknown field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: |
            The fence, the current session, the lease state, or the renewal
            generation rejected the renewal. Codes `fence_rejected`,
            `worker_session_unavailable`, `lease_expired`, `state_conflict`, or
            `renewal_conflict`. None of them mutates anything.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: The body exceeded the configured size limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: An identifier or the expected generation was invalid.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Sanitized internal error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: |
            A database call in this request failed because the request's own
            server-owned deadline elapsed, so the durable outcome is unknown to
            the caller. Code `service_unavailable`.

            The deadline can elapse while acquiring a lock, while executing a
            statement, or during COMMIT. A COMMIT that is cut short is
            ambiguous, so this response never asserts that nothing was
            committed.

            **Retry:** repeat the identical request with the same full fence —
            the `lease_id` in the path plus the same `job_id`, `attempt_id`,
            `worker_id`, and `worker_session_id` — and, critically, the same
            `renewal_request_id` and the same `expected_renewal_version`.
            Generating a new renewal request id or incrementing the expected
            version on a retry is what would extend the lease twice. Under the
            original pair the renewal is idempotent: if the first call
            committed, the replay returns that same window without moving the
            expiry again.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: service_unavailable
                  message: >-
                    the request exceeded its deadline before the durable outcome was
                    known; retry the identical request, which is safe because every
                    worker-control operation is idempotent under the identity and
                    fencing identifiers it already carries
...
  /internal/v1/attempts/{attempt_id}/start:
    post:
      tags: [Worker Control]
      summary: Fenced attempt start
      description: |
        Idempotently moves the exact current job and attempt `LEASED → RUNNING`.
        PostgreSQL server time must still be before the active lease expiry.
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FenceRequest' }
      responses:
        '204': { description: Started, or an exact already-started replay. }
        '400':
          description: Malformed JSON or an unknown field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Fence, state, current session, or lease expiry rejected the transition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: The body exceeded the configured size limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: An identifier was invalid.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Sanitized internal error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: |
            A database call in this request failed because the request's own
            server-owned deadline elapsed, so the durable outcome is unknown to
            the caller. Code `service_unavailable`.

            The deadline can elapse while acquiring a lock, while executing a
            statement, or during COMMIT. A COMMIT that is cut short is
            ambiguous, so this response never asserts that nothing was
            committed.

            **Retry:** repeat the identical request with the same attempt fence
            — the `attempt_id` in the path plus the same `job_id`, `lease_id`,
            `worker_id`, and `worker_session_id`. The start transition is
            idempotent under that fence: replaying it against an attempt that
            is already RUNNING succeeds without changing the recorded start
            time.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: service_unavailable
                  message: >-
                    the request exceeded its deadline before the durable outcome was
                    known; retry the identical request, which is safe because every
                    worker-control operation is idempotent under the identity and
                    fencing identifiers it already carries
...
components:
  schemas:
...
    HeartbeatRequest:
      type: object
      additionalProperties: false
      required: [worker_id]
      properties:
        worker_id: { type: string, format: uuid }
      description: |
        Deliberately carries no timestamp. PostgreSQL receipt time is the only
        authority for heartbeat staleness.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation_failed
                - malformed_json
                - unknown_queue
                - idempotency_conflict
                - not_found
                - payload_too_large
                - method_not_allowed
                - internal_error
                - service_unavailable
                - worker_session_conflict
                - worker_session_unavailable
                - claim_conflict
                - fence_rejected
                - lease_expired
                - state_conflict
                - renewal_conflict
              description: Stable machine-readable code. Branch on this, not on `message`.
            message: { type: string }
            details:
              type: array
              description: Field-level problems. Semantic validation reports all discovered field problems together.
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
            request_id: { type: string }